+++ CHANGELOG.md
@@ -9,7 +9,8 @@ All notable changes to this project will be documented in this file.
 
 - New `socket`, `socket_server` inputs.
 - New `socket` output.
-- Kafka connectors now support SASL using `OAUTHBEARER`.
+- Kafka connectors now support SASL using `OAUTHBEARER`, `SCRAM-SHA-256`,
+  `SCRAM-SHA-512` mechanisms.
 
 ### Changed